+++ lib/fetch_tray.dart
@@ -3,10 +3,14 @@ import 'package:http/http.dart' as http;
 export './contracts/tray_environment.dart';
 export './contracts/tray_request.dart';
 export './contracts/tray_request_body.dart';
+export './contracts/tray_request_metadata.dart';
 export './hooks/use_make_tray_request.dart';
 export './hooks/use_make_lazy_tray_request.dart';
 export './utils/make_tray_request.dart';
 export './utils/make_tray_testing_request.dart';
+export './interfaces/paginatable.dart';
+export './pagination_drivers/fetch_tray_pagination_driver.dart';
+export './pagination_drivers/page_pagination_driver.dart';
 
-// export httpclient to make it easier to import it
+// export http client to make it easier to import it
 typedef HttpClient = http.Client;